coreapi: split the interface into multiple files
License: MIT
