+++ core/coreapi/interface/util.go
@@ -1,0 +1,10 @@
+package iface
+
+import (
+	"io"
+)
+
+type Reader interface {
+	io.ReadSeeker
+	io.Closer
+}
